+++ app/models/post.rb
@@ -51,6 +51,13 @@ class Post < ApplicationRecord
 
   scope :all_public, -> { where(public: true) }
 
+  scope :all_local_public, -> {
+    where(" exists (
+      select 1 from people where posts.author_id = people.id
+      and people.pod_id is null)
+      and posts.public = true")
+  }
+
   scope :commented_by, ->(person)  {
     select('DISTINCT posts.*')
       .joins(:comments)